requirements.txt
langchain==1.2.15
langchain-core==1.3.0
langchain-community==0.4.1
langchain-groq==1.1.2
langchain-openai==1.1.14
langchain-ollama==1.1.0
langgraph==1.1.8
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.10
langgraph-sdk==0.3.13
fastembed==0.8.0
faiss-cpu==1.13.2
streamlit==1.56.0
python-dotenv==1.2.2
SQLAlchemy==2.0.49
pandas==3.0.2
pytest==9.0.3